services/shared-python/pyproject.toml
[project]
name = "infatium-shared"
version = "0.1.0"
description = "Shared Python code for Infatium services"
authors = [
    { name = "margosq", email = "slava1kvartovkin@gmail.com" }
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic-settings>=2.11.0",
    "nats-py>=2.7.0",
    "faststream[nats]>=0.6.4",
    "loguru>=0.7.3",
    "sentry-sdk>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "prometheus-client>=0.23.1",
    "redis>=5.0.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.6",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["."]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG001"]
ignore = ["E501", "B008", "C901", "W191"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
